backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
mysql-connector-python>=8.2.0
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
websockets>=12.0
apscheduler>=3.10.4
